requirements.txt
aiosqlite==0.19.0
anyio==3.6.2
click==8.1.3
colorama==0.4.6
databases==0.6.1
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
npm==0.1.1
npx==0.1.1
numpy==1.24.3
optional-django==0.1.0
ormar==0.12.0
psycopg2-binary==2.9.6
pydantic==1.10.2
PyJWT==2.7.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
